# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "any": {},
      "BaseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "requires": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "version": {
                  "type": "string"
                }
              },
              "required": [
                "code",
                "version"
              ]
            }
          },
          "entries": {
            "type": "object",
            "additionalProperties": {}
          },
          "content": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "name",
          "requires",
          "content"
        ]
      },
      "PagedResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "totalCount": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "totalCount"
        ]
      },
      "PagedResult<any>": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "totalCount": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "totalCount"
        ]
      },
      "PagedResult<PresetModel>": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "requires": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "version": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "code",
                      "version"
                    ]
                  }
                },
                "entries": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "content": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "code": {
                  "type": "string"
                },
                "version": {
                  "type": "string"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "code",
                "version",
                "tags"
              ]
            }
          },
          "totalCount": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "totalCount"
        ]
      },
      "PageOptions": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码，默认0"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数，默认20"
          },
          "search": {
            "description": "可选搜索项"
          }
        }
      },
      "PresetModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "requires": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "version": {
                  "type": "string"
                }
              },
              "required": [
                "code",
                "version"
              ]
            }
          },
          "entries": {
            "type": "object",
            "additionalProperties": {}
          },
          "content": {
            "type": "object",
            "additionalProperties": {}
          },
          "code": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "version",
          "tags"
        ]
      },
      "ReadableStream": {},
      "RequireModel": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "version"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                },
                "code": {
                  "type": "string",
                  "example": "404"
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource already exists"
                },
                "code": {
                  "type": "string",
                  "example": "409"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected errors occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/presets": {
      "get": {
        "operationId": "get-presets",
        "summary": "获取预设分页列表",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "number",
              "description": "页码，默认0"
            },
            "description": "页码，默认0"
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "number",
              "description": "每页条数，默认20"
            },
            "description": "每页条数，默认20"
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "description": "可选搜索项",
              "type": "string"
            },
            "description": "可选搜索项"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult<PresetModel>"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-presets",
        "summary": "创建预设",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/presets/{id}": {
      "get": {
        "operationId": "get-presets-{id}",
        "summary": "获取预设",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetModel"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "operationId": "put-presets-{id}",
        "summary": "更新预设",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/any"
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-presets-{id}",
        "summary": "删除预设",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/presets/{id}/export": {
      "get": {
        "operationId": "get-presets-{id}-export",
        "summary": "获取预设",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadableStream"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/presets/{id}/entries/{entryType}": {
      "get": {
        "operationId": "get-presets-{id}-entries-{entryType}",
        "summary": "获取条目分页列表",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "number",
              "description": "页码，默认0"
            },
            "description": "页码，默认0"
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "type": "number",
              "description": "每页条数，默认20"
            },
            "description": "每页条数，默认20"
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "description": "可选搜索项",
              "type": "string"
            },
            "description": "可选搜索项"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult<any>"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-presets-{id}-entries-{entryType}",
        "summary": "创建条目",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          },
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: entryType"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/any"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/presets/{id}/entries/{entryType}/{entryId}": {
      "put": {
        "operationId": "put-presets-{id}-entries-{entryType}-{entryId}",
        "summary": "更新条目",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          },
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: entryType"
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: entryId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/any"
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "operationId": "delete-presets-{id}-entries-{entryType}-{entryId}",
        "summary": "删除条目",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          },
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: entryType"
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: entryId"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/presets/{id}/entries/{entryType}/{entryId}/disabled": {
      "put": {
        "operationId": "put-presets-{id}-entries-{entryType}-{entryId}-disabled",
        "summary": "更新条目",
        "description": "",
        "tags": [
          "Presets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          },
          {
            "name": "entryType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example",
            "description": "Path parameter: entryType"
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: entryId"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Presets"
    }
  ]
}