# src/app/api/presets/[id]/entries/[entryType]/route.ts
// src/app/api/presets/route.ts
import {NextResponse} from 'next/server';
import {interceptor} from "@/server/interceptor";
import {presetRepository} from "@/server/business/presets";
import {like, or, SQL} from "drizzle-orm";
import {PageOptions} from "@/shared/models";

/**
 * 获取条目分页列表
 * @pathParams { id:string, entryType: string }
 * @params PageOptions
 * @response PagedResult<any>
 * @openapi
 */
export const GET = interceptor.createRoute(
    async (request, records) => {
        const {id, entryType} = await records.context.params as { id: string, entryType: string };
        const options = records.searchParams as PageOptions;
        const models = await presetRepository.entry.getList(id, entryType, options);
        return NextResponse.json(models);
    }
)

/**
 * 创建条目
 * @pathParams { id:string, entryType: string }
 * @body any
 * @response {id: number}
 * @openapi
 */
export const POST = interceptor.createRoute(
    async (request, records) => {
        const {id, entryType} = await records.context.params as { id: string, entryType: string };
        const model = records.body;
        const res = await presetRepository.entry.create(id, entryType, model);
        return NextResponse.json({id: res});
    }
)